requirements.txt
PyYAML>=6.0
matplotlib>=3.7.0
numpy>=1.24.0
llama-index>=0.10.0
llama-index-vector-stores-chroma
llama-index-embeddings-fastembed
fastembed
chromadb
python-dotenv
